Moved networking, modified viewcontroller
